Extract the OnePath specific code from Goal.
